refactor: move db json bindings to server_utils
We're going to use this in the auth server, so let's put it somewhere shared.
