+++ auth-server/auth_server/persistence/orm_models.py
@@ -2,11 +2,9 @@
 
 from __future__ import annotations
 
-import json
 from datetime import datetime
-from typing import Any, TypeAlias
 
-from sqlalchemy import ForeignKey, String, TypeDecorator, false
+from sqlalchemy import ForeignKey, false
 from sqlalchemy.orm import (
     DeclarativeBase,
     Mapped,
@@ -14,7 +12,7 @@ from sqlalchemy.orm import (
     relationship,
 )
 
-from server_utils.sql_utils import UTCDateTime
+from server_utils.sql_utils import JsonPythonValue, JsonValue, UTCDateTime
 
 
 class Base(DeclarativeBase):
@@ -73,36 +71,6 @@ class FailedLogin(Base):
     We don't use this for anything yet, but we're storing it in case we someday want to
     implement time-based lockouts or rate limiting.
     """
-
-
-JsonPythonValue: TypeAlias = (
-    str
-    | int
-    | float
-    | bool
-    | None
-    | list["JsonPythonValue"]
-    | dict[str, "JsonPythonValue"]
-)
-"""The output of `json.dumps()` / the input of `json.loads()`."""
-
-
-class JsonValue(TypeDecorator[object]):
-    """Transparently serializes Python values to/from JSON strings in the DB."""
-
-    impl = String
-    cache_ok = True
-
-    def process_bind_param(self, value: object | None, dialect: Any) -> str | None:
-        """Python → DB: json.dumps before writing."""
-        return json.dumps(value)
-
-    def process_result_value(self, value: str | None, dialect: Any) -> object | None:
-        """DB → Python: json.loads after reading."""
-        if value is not None:
-            result: object = json.loads(value)
-            return result
-        return None
 
 
 class Setting(Base):
